requirements.txt
Django==2.2.14
django-mdeditor==0.1.18
gunicorn==20.0.4
importlib-metadata==1.7.0
Markdown==3.2.2
mysqlclient==2.0.1
pytz==2020.1
sqlparse==0.3.1
zipp==3.1.0
Django==2.2.14
django-mdeditor==0.1.18
Faker==4.1.2
gunicorn==20.0.4
importlib-metadata==1.7.0
Markdown==3.2.2
mysqlclient==2.0.1
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
zipp==3.1.0
Django==2.2.14
django-mdeditor==0.1.18
django-pure-pagination==0.3.0
Faker==4.1.2
gunicorn==20.0.4
importlib-metadata==1.7.0
Markdown==3.2.2
mysqlclient==2.0.1
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
zipp==3.1.0
